feat(utils): add cuisines and diets data
Add a new file `cuisines.ts` to the `utils` directory, which exports an array of all the cuisines supported by the application.

Add a new file `diets.ts` to the `utils` directory, which exports an object containing information about all the diets supported by the application. Each diet object contains a name and an icon.